% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLCUB.R
\name{NLCUB}
\alias{NLCUB}
\title{Fitting NLCUB models}
\usage{
NLCUB(
  r,
  g = NULL,
  m = NULL,
  maxT = NULL,
  param0 = c(0.5, 0.5),
  freq.table = TRUE,
  method = "EM",
  draw.plot = TRUE,
  dk = NULL
)
}
\arguments{
\item{r}{numeric, observed ratings (either the vector of microdata or the vector of the m observed frequencies (frequency table)}

\item{g}{numeric vector, 'latent' categories assigned to each rating point; if \code{g} is declared, pai and xi are estimated for fixed g;if g is not declared, model selection is performed (see method) in order to determine optimal g}

\item{m}{integer, number of categories of the response scale    (active only when \code{g} is not declared)}

\item{maxT}{numeric, maximum value for T (must be maxT > m-1, default = 2m-1); active only when \code{g} is not declared}

\item{param0}{numeric, starting values for pai and xi (default: c(0.5,0.5))}

\item{freq.table}{logical, if TRUE, the data in r is the vector of the m observed frequencies (frequency table) (default=TRUE)}

\item{method}{character, "NM" (likelihood based - Melder-Mead maximization) - "EM" (likelihood based - EM algorithm)}

\item{draw.plot}{logical, if TRUE, graphs are plotted (default=TRUE)}

\item{dk}{numeric, proportion of 'don't know' responses; if declared, in addition to the estimate of pai, the estimated of pai adjusted for the presence of dk responses is provided}
}
\value{
A list with the following estimates:

\itemize{
\item parameter estimates (pai xi g)
}

\itemize{
\item fitted frequencies
}

\itemize{
\item diss index
}

\itemize{
\item average transition probabilities
}

\itemize{
\item transition probabilities matrix
}

\itemize{
\item unconditional transition probability
}

\itemize{
\item expected number of one-rating-point increments
}

The command can also displays two graphs: observed vs fitted frequencies + transition plot
}
\description{
Fitting NLCUB models
}
\examples{
N <- 1000
pai.sim <- 0.8
xi.sim <- 0.3
g.sim <- c(1,1,2,4,2)
cats <- 5
set.seed(1234567)
dataNLCUB <- simNLCUB(N, pai.sim, xi.sim, g.sim)
datitab <- as.matrix(table(dataNLCUB))
est <- NLCUB(datitab, g=g.sim, freq.table=TRUE)
}
\references{
M. Manisera and P.Zuccolotto (2014) Modeling rating data with Nonlinear CUB models. Computational Statistics and Data Analysis, 78, pp. 100–118

M. Manisera and P.Zuccolotto (2014) Nonlinear CUB models: The R code. Statistical Software - Statistica & Applicazioni, Vol. XII, n. 2, pp. 205-223
}
\author{
Paola Zuccolotto, Marica Manisera, Sandri Marco
}
