% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLCUBplot.R
\name{NLCUBplot}
\alias{NLCUBplot}
\title{Plot observed relative frequencies nad fitted probabilities for different ratings}
\usage{
NLCUBplot(rp1, paip1, xip1, gp1, freq.table.p1 = TRUE)
}
\arguments{
\item{rp1}{numeric observed ratings (either the vector of microdata or the vector of the m observed frequencies (frequency table)}

\item{paip1}{numeric, value of the pai parameter}

\item{xip1}{numeric, value of the csi parameter}

\item{gp1}{numeric vector, 'latent' categories assigned to each rating point}

\item{freq.table.p1}{logical, if TRUE, the data in r is the vector of the m observed frequencies (frequency table)}
}
\value{
A ggplot object
}
\description{
Plot observed relative frequencies nad fitted probabilities for different ratings
}
\details{
(Details here)
}
\examples{
N <- 1000
pai.sim <- 0.8
xi.sim <- 0.3
g.sim <- c(1,1,2,4,2)
cats <- 5
set.seed(1234567)
}
\references{
M. Manisera and P. Zuccolotto (2014) Modeling rating data with Nonlinear CUB models. Computational Statistics and Data Analysis, 78, pp. 100–118

M. Manisera and P. Zuccolotto (2014) Nonlinear CUB models: The R code. Statistical Software - Statistica & Applicazioni, Vol. XII, n. 2, pp. 205-223
}
\author{
Paola Zuccolotto, Marica Manisera, Sandri Marco
}
