% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.NLCUB.R
\name{plot.NLCUB}
\alias{plot.NLCUB}
\title{Plots a NLCUB model from a 'NLCUB' object}
\usage{
\method{plot}{NLCUB}(x, log.scale = TRUE, freq.table.p1 = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{NLCUB}}

\item{log.scale}{logical, if TRUE, the logarithmic scale is used; if FALSE, the linear scale is used}

\item{freq.table.p1}{logical, if TRUE, the data in r is the vector of the m observed frequencies (frequency table)}

\item{...}{other graphical parameters.}
}
\value{
A ggplot object
}
\description{
Plots a NLCUB model from a 'NLCUB' object
}
\details{
(Details here)
}
\examples{
N <- 1000
pai.sim <- 0.8
xi.sim <- 0.3
g.sim <- c(1,1,2,4,2)
cats <- 5
set.seed(1234567)
dataNLCUB <- simNLCUB(N, pai.sim, xi.sim, g.sim)
datitab <- table(dataNLCUB)
est <- NLCUB(datitab, g=g.sim, freq.table=TRUE)
plot(est)
}
\references{
M. Manisera and P. Zuccolotto (2014) Modeling rating data with Nonlinear CUB models. Computational Statistics and Data Analysis, 78, pp. 100–118

M. Manisera and P. Zuccolotto (2014) Nonlinear CUB models: The R code. Statistical Software - Statistica & Applicazioni, Vol. XII, n. 2, pp. 205-223
}
\author{
Paola Zuccolotto, Marica Manisera, Sandri Marco
}
