% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transplot.R
\name{Transplot}
\alias{Transplot}
\title{Transition plot}
\usage{
Transplot(xip2, gp2, log.scale = TRUE)
}
\arguments{
\item{xip2}{numeric, value of the csi parameter}

\item{gp2}{numeric vector, 'latent' categories assigned to each rating point}

\item{log.scale}{logical, if TRUE, the logarithmic scale is used; if FALSE, the linear scale is used}
}
\value{
A ggplot object
}
\description{
Transition plot
}
\details{
(Details here)
}
\examples{
N <- 1000
pai.sim <- 0.8
xi.sim <- 0.3
g.sim <- c(1,1,2,4,2)
cats <- 5
set.seed(1234567)
}
\references{
M. Manisera and P. Zuccolotto (2014) Modeling rating data with Nonlinear CUB models. Computational Statistics and Data Analysis, 78, pp. 100–118

M. Manisera and P. Zuccolotto (2014) Nonlinear CUB models: The R code. Statistical Software - Statistica & Applicazioni, Vol. XII, n. 2, pp. 205-223
}
\author{
Paola Zuccolotto, Marica Manisera, Sandri Marco
}
