% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simNLCUB.R
\name{simNLCUB}
\alias{simNLCUB}
\title{Computes, for each action, an estimate of the value of the shotclock when the action has ended}
\usage{
simNLCUB(Nsim, paisim, xisim, gsim)
}
\arguments{
\item{Nsim}{integer, sample size}

\item{paisim}{numeric, value of the \code{pai} parameter to be used for simulation}

\item{xisim}{numeric, value of the \code{csi} parameter to be used for simulation}

\item{gsim}{vector of the 'latent' categories assigned to each rating point}
}
\value{
A data frame with N ratings simulated according to a NLCUB model with parameters \code{paisim}, \code{csisim}, \code{gsim}
}
\description{
Computes, for each action, an estimate of the value of the shotclock when the action has ended
}
\details{
It is necessary that the name of the team is contained in the column corresponding to the description
}
\examples{
N <- 1000
pai.sim <- 0.8
xi.sim <- 0.3
g.sim <- c(1,1,2,4,2)
cats <- 5
set.seed(1234567)
dataNLCUB <- simNLCUB(N,pai.sim, xi.sim, g.sim)
}
\references{
M. Manisera and P.Zuccolotto (2014) Modeling rating data with Nonlinear CUB models. Computational Statistics and Data Analysis, 78, pp. 100–118

M. Manisera and P.Zuccolotto (2014) Nonlinear CUB models: The R code. Statistical Software - Statistica & Applicazioni, Vol. XII, n. 2, pp. 205-223
}
\author{
Paola Zuccolotto, Marica Manisera, Sandri Marco
}
